refactor: extract 10 handler modules from api/mod.rs (8168→5353 lines)
Extract handler groups into focused submodules:
- absences.rs: CRUD + patterns + team absences (460 lines)
- announcements.rs: public + admin CRUD (317 lines)
- calendar.rs: combined events view + iCal export (248 lines)
- guest.rs: guest bookings + admin management (247 lines)
- notifications.rs: list + mark read (153 lines)
- recurring.rs: recurring booking CRUD (302 lines)
- settings.rs: admin settings + feature flags + themes (524 lines)
- swap.rs: slot swap requests (359 lines)
- team.rs: team status today + member list (141 lines)
- waitlist.rs: lot waitlist CRUD (181 lines)

All 579 tests pass. Zero clippy warnings. OpenAPI paths updated.
Closes #101, closes #61
