+++ parkhub-server/src/api/mod.rs
@@ -47,10 +47,6 @@ pub const MAX_PHOTO_BYTES: usize = 2 * 1024 * 1024;
 /// German standard VAT rate (19% — Umsatzsteuergesetz § 12 Abs. 1)
 const VAT_RATE: f64 = 0.19;
 
-use parkhub_common::models::{
-    Absence, AbsencePattern, AbsenceType, Announcement, AnnouncementSeverity, GuestBooking,
-    Notification, RecurringBooking, SwapRequest, SwapRequestStatus, WaitlistEntry,
-};
 use parkhub_common::{
     ApiResponse, Booking, BookingPricing, BookingStatus, CreateBookingRequest, CreditTransaction,
     CreditTransactionType, HandshakeRequest, HandshakeResponse, LoginResponse, PaymentStatus,
@@ -66,43 +62,74 @@ type SharedState = Arc<RwLock<AppState>>;
 // Submodules
 // ─────────────────────────────────────────────────────────────────────────────
 
+pub mod absences;
 pub mod admin;
+pub mod announcements;
 pub mod auth;
 mod bookings;
 pub mod branding;
+pub mod calendar;
 pub mod credits;
 pub mod export;
 pub mod favorites;
+pub mod guest;
 pub mod import;
 pub mod lots;
+pub mod notifications;
 pub mod payments;
 pub mod push;
 pub mod pwa;
 pub mod qr;
 pub mod recommendations;
+pub mod recurring;
+pub mod settings;
 pub mod setup;
 mod social;
+pub mod swap;
+pub mod team;
 pub mod translations;
 mod users;
 pub mod vehicles;
+pub mod waitlist;
 pub mod webhooks;
 pub mod ws;
 pub mod zones;
 
 // Re-import handler functions so the router can reference them unqualified.
+use absences::{
+    create_absence, delete_absence, get_absence_pattern, list_absences, list_team_absences,
+    save_absence_pattern, update_absence,
+};
+use announcements::{
+    admin_create_announcement, admin_delete_announcement, admin_list_announcements,
+    admin_update_announcement, get_active_announcements,
+};
 use auth::{forgot_password, login, refresh_token, register, reset_password};
+use calendar::{calendar_events, user_calendar_ics};
 use credits::{
     admin_grant_credits, admin_list_credit_transactions, admin_refill_all_credits,
     admin_update_user_quota, get_user_credits,
 };
 use export::{admin_export_bookings_csv, admin_export_revenue_csv, admin_export_users_csv};
 use favorites::{add_favorite, list_favorites, remove_favorite};
+use guest::{admin_cancel_guest_booking, admin_list_guest_bookings, create_guest_booking};
 use import::import_users_csv;
 use lots::{
     create_lot, create_slot, delete_lot, delete_slot, get_lot, get_lot_pricing, get_lot_slots,
     list_lots, update_lot, update_lot_pricing, update_slot,
 };
+use notifications::{list_notifications, mark_all_notifications_read, mark_notification_read};
 use recommendations::get_recommendations;
+use recurring::{
+    create_recurring_booking, delete_recurring_booking, list_recurring_bookings,
+    update_recurring_booking,
+};
+use settings::{
+    admin_get_features, admin_get_settings, admin_get_use_case, admin_update_features,
+    admin_update_settings, get_features, get_public_theme, read_admin_setting,
+};
+use swap::{create_swap_request, list_swap_requests, update_swap_request};
+use team::{team_list, team_today};
 use translations::{
     create_proposal, get_proposal, list_overrides, list_proposals, review_proposal,
     vote_on_proposal,
@@ -111,6 +138,7 @@ use vehicles::{
     create_vehicle, delete_vehicle, get_vehicle_photo, list_vehicles, update_vehicle,
     upload_vehicle_photo, vehicle_city_codes,
 };
+use waitlist::{join_waitlist, leave_waitlist, list_waitlist};
 use webhooks::{create_webhook, delete_webhook, list_webhooks, test_webhook, update_webhook};
 use zones::{create_zone, delete_zone, list_zones, update_zone};
 
@@ -3369,2098 +3397,6 @@ pub async fn admin_list_bookings(
 }
 
 // ═══════════════════════════════════════════════════════════════════════════════
-// ADMIN SETTINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// All admin settings with their default values.
-const ADMIN_SETTINGS: &[(&str, &str)] = &[
-    ("company_name", "ParkHub"),
-    ("use_case", "company"),
-    ("self_registration", "true"),
-    ("license_plate_mode", "optional"),
-    ("display_name_format", "first_name"),
-    ("max_bookings_per_day", "0"),
-    ("allow_guest_bookings", "false"),
-    ("auto_release_enabled", "false"),
-    ("auto_release_minutes", "30"),
-    ("require_vehicle", "false"),
-    ("waitlist_enabled", "true"),
-    ("min_booking_duration_hours", "0"),
-    ("max_booking_duration_hours", "0"),
-    ("credits_enabled", "false"),
-    ("credits_per_booking", "1"),
-];
-
-/// Read a single admin setting from DB, falling back to its default.
-async fn read_admin_setting(db: &crate::db::Database, key: &str) -> String {
-    if let Ok(Some(val)) = db.get_setting(key).await {
-        return val;
-    }
-    ADMIN_SETTINGS
-        .iter()
-        .find(|(k, _)| *k == key)
-        .map(|(_, v)| v.to_string())
-        .unwrap_or_default()
-}
-
-/// Use-case theme definitions — maps `use_case` key to display config
-fn use_case_theme(key: &str) -> serde_json::Value {
-    match key {
-        "company" => serde_json::json!({
-            "key": "company",
-            "name": "Company Parking",
-            "description": "Employee parking for offices and campuses",
-            "icon": "buildings",
-            "primary_color": "#0d9488",
-            "accent_color": "#0ea5e9",
-            "terminology": {
-                "user": "Employee", "users": "Employees",
-                "lot": "Parking Area", "slot": "Spot",
-                "booking": "Reservation", "department": "Department"
-            },
-            "features_emphasis": ["team_calendar", "absence_tracking", "departments", "credits"]
-        }),
-        "residential" => serde_json::json!({
-            "key": "residential",
-            "name": "Residential Parking",
-            "description": "Parking for apartment buildings and housing complexes",
-            "icon": "house-line",
-            "primary_color": "#059669",
-            "accent_color": "#84cc16",
-            "terminology": {
-                "user": "Resident", "users": "Residents",
-                "lot": "Parking Area", "slot": "Space",
-                "booking": "Reservation", "department": "Unit"
-            },
-            "features_emphasis": ["guest_parking", "long_term_bookings", "public_display"]
-        }),
-        "shared" => serde_json::json!({
-            "key": "shared",
-            "name": "Shared Parking",
-            "description": "Community or co-working parking spaces",
-            "icon": "users-three",
-            "primary_color": "#7c3aed",
-            "accent_color": "#06b6d4",
-            "terminology": {
-                "user": "Member", "users": "Members",
-                "lot": "Parking Zone", "slot": "Spot",
-                "booking": "Booking", "department": "Group"
-            },
-            "features_emphasis": ["quick_book", "waitlist", "public_display", "qr_codes"]
-        }),
-        "rental" => serde_json::json!({
-            "key": "rental",
-            "name": "Rental / Commercial",
-            "description": "Paid parking for customers and tenants",
-            "icon": "currency-circle-dollar",
-            "primary_color": "#2563eb",
-            "accent_color": "#f59e0b",
-            "terminology": {
-                "user": "Customer", "users": "Customers",
-                "lot": "Parking Facility", "slot": "Bay",
-                "booking": "Rental", "department": "Account"
-            },
-            "features_emphasis": ["invoicing", "pricing", "revenue_reports", "guest_bookings"]
-        }),
-        _ => serde_json::json!({
-            "key": "personal",
-            "name": "Personal / Private",
-            "description": "Private parking for family and friends",
-            "icon": "car-simple",
-            "primary_color": "#e11d48",
-            "accent_color": "#f97316",
-            "terminology": {
-                "user": "Person", "users": "People",
-                "lot": "Driveway", "slot": "Spot",
-                "booking": "Booking", "department": "Group"
-            },
-            "features_emphasis": ["simple_booking", "guest_parking"]
-        }),
-    }
-}
-
-/// `GET /api/v1/admin/settings/use-case` — return current use-case with theme config
-#[utoipa::path(
-    get,
-    path = "/api/v1/admin/settings/use-case",
-    tag = "Admin",
-    summary = "Get use-case configuration",
-    description = "Return current use-case with theme config. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_get_use_case(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-    let current = read_admin_setting(&state_guard.db, "use_case").await;
-    let theme = use_case_theme(&current);
-    let all_options: Vec<serde_json::Value> =
-        ["company", "residential", "shared", "rental", "personal"]
-            .iter()
-            .map(|k| use_case_theme(k))
-            .collect();
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "current": theme,
-            "available": all_options,
-        }))),
-    )
-}
-
-/// `GET /api/v1/admin/settings` — return all settings (merged defaults + stored values)
-#[utoipa::path(get, path = "/api/v1/admin/settings", tag = "Admin",
-    summary = "Get system settings (admin)", description = "Returns all system settings. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Settings"), (status = 403, description = "Forbidden"))
-)]
-pub async fn admin_get_settings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let mut data = serde_json::Map::new();
-    for (key, default_val) in ADMIN_SETTINGS {
-        let value = state_guard
-            .db
-            .get_setting(key)
-            .await
-            .ok()
-            .flatten()
-            .unwrap_or_else(|| default_val.to_string());
-        data.insert(key.to_string(), serde_json::Value::String(value));
-    }
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::Value::Object(data))),
-    )
-}
-
-/// Validate a settings value against its allowed options.
-fn validate_setting_value(key: &str, value: &str) -> Result<(), &'static str> {
-    match key {
-        "use_case" => {
-            if !["company", "residential", "shared", "rental", "personal"].contains(&value) {
-                return Err("use_case must be company, residential, shared, rental, or personal");
-            }
-        }
-        "self_registration"
-        | "allow_guest_bookings"
-        | "require_vehicle"
-        | "waitlist_enabled"
-        | "credits_enabled"
-        | "auto_release_enabled" => {
-            if value != "true" && value != "false" {
-                return Err("Value must be \"true\" or \"false\"");
-            }
-        }
-        "license_plate_mode" => {
-            if !["required", "optional", "disabled"].contains(&value) {
-                return Err("license_plate_mode must be required, optional, or disabled");
-            }
-        }
-        "display_name_format" => {
-            if !["first_name", "full_name", "username"].contains(&value) {
-                return Err("display_name_format must be first_name, full_name, or username");
-            }
-        }
-        "max_bookings_per_day" | "auto_release_minutes" | "credits_per_booking" => {
-            if value.parse::<i32>().is_err() {
-                return Err("Value must be an integer");
-            }
-        }
-        "min_booking_duration_hours" | "max_booking_duration_hours" => {
-            if value.parse::<f64>().is_err() {
-                return Err("Value must be a number");
-            }
-        }
-        "company_name" => { /* any string is fine */ }
-        _ => return Err("Unknown setting key"),
-    }
-    Ok(())
-}
-
-/// `PUT /api/v1/admin/settings` — update one or more settings (admin only)
-#[utoipa::path(put, path = "/api/v1/admin/settings", tag = "Admin",
-    summary = "Update system settings (admin)", description = "Saves system settings. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Saved"), (status = 403, description = "Forbidden"))
-)]
-pub async fn admin_update_settings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(payload): Json<serde_json::Value>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let Some(obj) = payload.as_object() else {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_INPUT",
-                "Request body must be a JSON object of key-value pairs",
-            )),
-        );
-    };
-
-    let allowed_keys: Vec<&str> = ADMIN_SETTINGS.iter().map(|(k, _)| *k).collect();
-    let mut updated = serde_json::Map::new();
-
-    for (key, val) in obj {
-        if !allowed_keys.contains(&key.as_str()) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_KEY",
-                    format!("Unknown setting: {key}"),
-                )),
-            );
-        }
-
-        let value_str = val.as_str().map_or_else(
-            || val.to_string().trim_matches('"').to_string(),
-            String::from,
-        );
-
-        if let Err(msg) = validate_setting_value(key, &value_str) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error("VALIDATION_ERROR", msg)),
-            );
-        }
-
-        if let Err(e) = state_guard.db.set_setting(key, &value_str).await {
-            tracing::error!("Failed to save setting {}: {}", key, e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to save setting")),
-            );
-        }
-
-        updated.insert(key.clone(), serde_json::Value::String(value_str));
-    }
-
-    // Audit log
-    if state_guard.config.audit_logging_enabled {
-        let _entry = AuditEntry::new(AuditEventType::ConfigChanged)
-            .user(auth_user.user_id, "admin")
-            .resource("settings", "admin_settings")
-            .details(serde_json::json!({ "updated": updated }))
-            .log();
-    }
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::Value::Object(updated))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// FEATURE FLAGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// All available feature module IDs.
-const FEATURE_MODULES: &[&str] = &[
-    "credits",
-    "absences",
-    "vehicles",
-    "analytics",
-    "team_view",
-    "booking_types",
-    "invoices",
-    "self_registration",
-    "generative_bg",
-    "micro_animations",
-    "fab_quick_actions",
-    "rich_empty_states",
-    "onboarding_hints",
-];
-
-/// Default enabled features (business use case).
-const DEFAULT_FEATURES: &[&str] = &[
-    "credits",
-    "absences",
-    "vehicles",
-    "analytics",
-    "team_view",
-    "booking_types",
-    "invoices",
-    "generative_bg",
-    "micro_animations",
-    "fab_quick_actions",
-    "rich_empty_states",
-    "onboarding_hints",
-];
-
-const SETTINGS_FEATURES_KEY: &str = "features_enabled";
-
-/// Read enabled features from DB, falling back to defaults.
-async fn read_features(db: &crate::db::Database) -> Vec<String> {
-    match db.get_setting(SETTINGS_FEATURES_KEY).await {
-        Ok(Some(json_str)) => serde_json::from_str::<Vec<String>>(&json_str).unwrap_or_else(|_| {
-            DEFAULT_FEATURES
-                .iter()
-                .map(std::string::ToString::to_string)
-                .collect()
-        }),
-        _ => DEFAULT_FEATURES
-            .iter()
-            .map(std::string::ToString::to_string)
-            .collect(),
-    }
-}
-
-/// `GET /api/v1/features` — public endpoint returning enabled features
-#[utoipa::path(get, path = "/api/v1/features", tag = "Public",
-    summary = "Get enabled feature flags",
-    description = "Returns enabled and available features. No auth required.",
-    responses((status = 200, description = "Success"))
-)]
-pub async fn get_features(
-    State(state): State<SharedState>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    let enabled = read_features(&state_guard.db).await;
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "enabled": enabled,
-            "available": FEATURE_MODULES,
-        }))),
-    )
-}
-
-/// `GET /api/v1/theme` — public: return current use-case theme (no auth required)
-#[utoipa::path(get, path = "/api/v1/theme", tag = "Public",
-    summary = "Get current theme",
-    description = "Returns theme and company name. No auth required.",
-    responses((status = 200, description = "Success"))
-)]
-pub async fn get_public_theme(
-    State(state): State<SharedState>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    let use_case = read_admin_setting(&state_guard.db, "use_case").await;
-    let company = read_admin_setting(&state_guard.db, "company_name").await;
-    let theme = use_case_theme(&use_case);
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "use_case": theme,
-            "company_name": company,
-        }))),
-    )
-}
-
-/// `GET /api/v1/admin/features` — admin: get features with full metadata
-#[utoipa::path(get, path = "/api/v1/admin/features", tag = "Admin",
-    summary = "Get feature flags (admin)",
-    description = "Returns feature modules with status. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_get_features(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let enabled = read_features(&state_guard.db).await;
-
-    let available: Vec<serde_json::Value> = FEATURE_MODULES
-        .iter()
-        .map(|id| {
-            serde_json::json!({
-                "id": id,
-                "enabled": enabled.contains(&id.to_string()),
-                "default_enabled": DEFAULT_FEATURES.contains(id),
-            })
-        })
-        .collect();
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "enabled": enabled,
-            "available": available,
-        }))),
-    )
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-pub struct UpdateFeaturesRequest {
-    enabled: Vec<String>,
-}
-
-/// `PUT /api/v1/admin/features` — admin: update enabled features
-#[utoipa::path(put, path = "/api/v1/admin/features", tag = "Admin",
-    summary = "Update feature flags (admin)",
-    description = "Sets enabled feature modules. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_update_features(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(body): Json<UpdateFeaturesRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let state_guard = state.write().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    // Validate: only accept known feature IDs
-    let valid: Vec<String> = body
-        .enabled
-        .iter()
-        .filter(|id| FEATURE_MODULES.contains(&id.as_str()))
-        .cloned()
-        .collect();
-
-    let json_str = serde_json::to_string(&valid).unwrap_or_default();
-    if let Err(e) = state_guard
-        .db
-        .set_setting(SETTINGS_FEATURES_KEY, &json_str)
-        .await
-    {
-        tracing::error!("Failed to save feature flags: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to save features",
-            )),
-        );
-    }
-
-    // Audit log
-    if state_guard.config.audit_logging_enabled {
-        let _entry = AuditEntry::new(AuditEventType::ConfigChanged)
-            .user(auth_user.user_id, "admin")
-            .resource("settings", "features_enabled")
-            .details(serde_json::json!({ "features": valid }))
-            .log();
-    }
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "enabled": valid,
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ABSENCES
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[derive(Deserialize)]
-pub struct AbsenceQuery {
-    #[serde(rename = "type")]
-    absence_type: Option<AbsenceType>,
-}
-
-/// `GET /api/v1/absences` — list current user's absences, optionally filtered by type
-#[utoipa::path(get, path = "/api/v1/absences", tag = "Absences",
-    summary = "List user absences",
-    description = "Returns absences for the authenticated user.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn list_absences(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Query(query): Query<AbsenceQuery>,
-) -> (StatusCode, Json<ApiResponse<Vec<Absence>>>) {
-    let state_guard = state.read().await;
-    match state_guard
-        .db
-        .list_absences_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(absences) => {
-            let filtered = match query.absence_type {
-                Some(ref t) => absences
-                    .into_iter()
-                    .filter(|a| &a.absence_type == t)
-                    .collect(),
-                None => absences,
-            };
-            (StatusCode::OK, Json(ApiResponse::success(filtered)))
-        }
-        Err(e) => {
-            tracing::error!("Failed to list absences: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list absences",
-                )),
-            )
-        }
-    }
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-pub struct CreateAbsenceRequest {
-    absence_type: AbsenceType,
-    start_date: String,
-    end_date: String,
-    note: Option<String>,
-}
-
-/// Validate a date string is YYYY-MM-DD format.
-fn is_valid_date(s: &str) -> bool {
-    chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").is_ok()
-}
-
-/// `POST /api/v1/absences` — create an absence
-#[utoipa::path(post, path = "/api/v1/absences", tag = "Absences",
-    summary = "Create an absence",
-    description = "Records a new absence for the authenticated user.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn create_absence(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<CreateAbsenceRequest>,
-) -> (StatusCode, Json<ApiResponse<Absence>>) {
-    if !is_valid_date(&req.start_date) || !is_valid_date(&req.end_date) {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_INPUT",
-                "Dates must be in YYYY-MM-DD format",
-            )),
-        );
-    }
-
-    if req.start_date > req.end_date {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_INPUT",
-                "start_date must not be after end_date",
-            )),
-        );
-    }
-
-    let absence = Absence {
-        id: Uuid::new_v4(),
-        user_id: auth_user.user_id,
-        absence_type: req.absence_type,
-        start_date: req.start_date,
-        end_date: req.end_date,
-        note: req.note,
-        source: "manual".to_string(),
-        created_at: Utc::now(),
-    };
-
-    let state_guard = state.read().await;
-    match state_guard.db.save_absence(&absence).await {
-        Ok(()) => (StatusCode::CREATED, Json(ApiResponse::success(absence))),
-        Err(e) => {
-            tracing::error!("Failed to save absence: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to create absence",
-                )),
-            )
-        }
-    }
-}
-
-/// `DELETE /api/v1/absences/{id}` — delete own absence
-#[utoipa::path(delete, path = "/api/v1/absences/{id}", tag = "Absences",
-    summary = "Delete an absence",
-    description = "Removes an absence owned by the user.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn delete_absence(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-
-    // Verify ownership
-    let absence = match state_guard.db.get_absence(&id).await {
-        Ok(Some(a)) => a,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Absence not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error fetching absence: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    if absence.user_id != auth_user.user_id {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    }
-
-    match state_guard.db.delete_absence(&id).await {
-        Ok(true) => (StatusCode::OK, Json(ApiResponse::success(()))),
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Absence not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to delete absence: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to delete absence",
-                )),
-            )
-        }
-    }
-}
-
-/// `GET /api/v1/absences/team` — list all team absences
-#[utoipa::path(
-    get,
-    path = "/api/v1/absences/team",
-    tag = "Absences",
-    summary = "List team absences",
-    description = "List all team member absences visible to the current user.",
-    security(("bearer_auth" = []))
-)]
-pub async fn list_team_absences(
-    State(state): State<SharedState>,
-    Extension(_auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<Absence>>>) {
-    let state_guard = state.read().await;
-    match state_guard.db.list_absences_team().await {
-        Ok(absences) => (StatusCode::OK, Json(ApiResponse::success(absences))),
-        Err(e) => {
-            tracing::error!("Failed to list team absences: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list team absences",
-                )),
-            )
-        }
-    }
-}
-
-/// `GET /api/v1/absences/pattern` — get user's absence pattern
-#[utoipa::path(
-    get,
-    path = "/api/v1/absences/pattern",
-    tag = "Absences",
-    summary = "Get absence pattern",
-    description = "Get the current user's recurring absence pattern.",
-    security(("bearer_auth" = []))
-)]
-pub async fn get_absence_pattern(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Option<AbsencePattern>>>) {
-    let state_guard = state.read().await;
-    let key = format!("absence_pattern:{}", auth_user.user_id);
-    match state_guard.db.get_setting(&key).await {
-        Ok(val) => {
-            let pattern =
-                val.and_then(|json_str| serde_json::from_str::<AbsencePattern>(&json_str).ok());
-            (StatusCode::OK, Json(ApiResponse::success(pattern)))
-        }
-        Err(e) => {
-            tracing::error!("Failed to get absence pattern: {e}");
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to get absence pattern",
-                )),
-            )
-        }
-    }
-}
-
-/// `POST /api/v1/absences/pattern` — save user's absence pattern
-#[utoipa::path(
-    post,
-    path = "/api/v1/absences/pattern",
-    tag = "Absences",
-    summary = "Save absence pattern",
-    description = "Save or update the current user's recurring absence pattern (e.g. homeoffice every Monday).",
-    security(("bearer_auth" = []))
-)]
-pub async fn save_absence_pattern(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(pattern): Json<AbsencePattern>,
-) -> (StatusCode, Json<ApiResponse<AbsencePattern>>) {
-    let state_guard = state.read().await;
-    let key = format!("absence_pattern:{}", auth_user.user_id);
-    let json_str = match serde_json::to_string(&pattern) {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!("Failed to serialize absence pattern: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Serialization error")),
-            );
-        }
-    };
-
-    match state_guard.db.set_setting(&key, &json_str).await {
-        Ok(()) => (StatusCode::OK, Json(ApiResponse::success(pattern))),
-        Err(e) => {
-            tracing::error!("Failed to save absence pattern: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to save absence pattern",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// TEAM VIEW
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[derive(Serialize)]
-pub struct TeamMemberStatus {
-    user_id: Uuid,
-    name: String,
-    username: String,
-    status: String,
-    absence_type: Option<AbsenceType>,
-}
-
-/// `GET /api/v1/team/today` — return all users with their status today
-#[utoipa::path(get, path = "/api/v1/team/today", tag = "Team",
-    summary = "Team status today",
-    description = "Returns all users with their status for today.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn team_today(
-    State(state): State<SharedState>,
-    Extension(_auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<TeamMemberStatus>>>) {
-    let state_guard = state.read().await;
-
-    let users = match state_guard.db.list_users().await {
-        Ok(u) => u,
-        Err(e) => {
-            tracing::error!("Failed to list users: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Failed to load users")),
-            );
-        }
-    };
-
-    let today = Utc::now().format("%Y-%m-%d").to_string();
-
-    let absences = state_guard
-        .db
-        .list_absences_team()
-        .await
-        .unwrap_or_default();
-
-    let bookings = state_guard.db.list_bookings().await.unwrap_or_default();
-
-    let mut result = Vec::new();
-    for user in &users {
-        if !user.is_active {
-            continue;
-        }
-
-        // Check for absence today
-        let user_absence = absences
-            .iter()
-            .find(|a| a.user_id == user.id && a.start_date <= today && a.end_date >= today);
-
-        if let Some(absence) = user_absence {
-            let status = match absence.absence_type {
-                AbsenceType::Homeoffice => "homeoffice",
-                AbsenceType::Vacation => "vacation",
-                AbsenceType::Sick => "sick",
-                AbsenceType::Training | AbsenceType::Other => "absent",
-            };
-            result.push(TeamMemberStatus {
-                user_id: user.id,
-                name: user.name.clone(),
-                username: user.username.clone(),
-                status: status.to_string(),
-                absence_type: Some(absence.absence_type.clone()),
-            });
-            continue;
-        }
-
-        // Check for booking today (confirmed or active)
-        let has_booking = bookings.iter().any(|b| {
-            b.user_id == user.id
-                && (b.status == BookingStatus::Confirmed || b.status == BookingStatus::Active)
-                && b.start_time.format("%Y-%m-%d").to_string() <= today
-                && b.end_time.format("%Y-%m-%d").to_string() >= today
-        });
-
-        let status = if has_booking { "parked" } else { "available" };
-        result.push(TeamMemberStatus {
-            user_id: user.id,
-            name: user.name.clone(),
-            username: user.username.clone(),
-            status: status.to_string(),
-            absence_type: None,
-        });
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(result)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ANNOUNCEMENTS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/announcements/active` — public, return active non-expired announcements
-#[utoipa::path(get, path = "/api/v1/announcements/active", tag = "Public",
-    summary = "Get active announcements",
-    description = "Returns active non-expired announcements. No auth required.",
-    responses((status = 200, description = "Success"))
-)]
-pub async fn get_active_announcements(
-    State(state): State<SharedState>,
-) -> (StatusCode, Json<ApiResponse<Vec<Announcement>>>) {
-    let state_guard = state.read().await;
-    match state_guard.db.list_announcements().await {
-        Ok(announcements) => {
-            let now = Utc::now();
-            let active: Vec<Announcement> = announcements
-                .into_iter()
-                .filter(|a| a.active && a.expires_at.is_none_or(|exp| exp > now))
-                .collect();
-            (StatusCode::OK, Json(ApiResponse::success(active)))
-        }
-        Err(e) => {
-            tracing::error!("Failed to list announcements: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list announcements",
-                )),
-            )
-        }
-    }
-}
-
-/// `GET /api/v1/admin/announcements` — admin: list all announcements
-#[utoipa::path(get, path = "/api/v1/admin/announcements", tag = "Admin",
-    summary = "List all announcements (admin)",
-    description = "Returns all announcements. Admin only.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn admin_list_announcements(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<Announcement>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    match state_guard.db.list_announcements().await {
-        Ok(announcements) => (StatusCode::OK, Json(ApiResponse::success(announcements))),
-        Err(e) => {
-            tracing::error!("Failed to list announcements: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list announcements",
-                )),
-            )
-        }
-    }
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-pub struct CreateAnnouncementRequest {
-    title: String,
-    message: String,
-    severity: AnnouncementSeverity,
-    active: Option<bool>,
-    expires_at: Option<DateTime<Utc>>,
-}
-
-/// `POST /api/v1/admin/announcements` — admin: create announcement
-#[utoipa::path(
-    post,
-    path = "/api/v1/admin/announcements",
-    tag = "Admin",
-    summary = "Create announcement",
-    description = "Create a new system announcement. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_create_announcement(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<CreateAnnouncementRequest>,
-) -> (StatusCode, Json<ApiResponse<Announcement>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let announcement = Announcement {
-        id: Uuid::new_v4(),
-        title: req.title,
-        message: req.message,
-        severity: req.severity,
-        active: req.active.unwrap_or(true),
-        created_by: Some(auth_user.user_id),
-        expires_at: req.expires_at,
-        created_at: Utc::now(),
-    };
-
-    match state_guard.db.save_announcement(&announcement).await {
-        Ok(()) => {
-            let audit = AuditEntry::new(AuditEventType::ConfigChanged)
-                .user(auth_user.user_id, "admin")
-                .resource("announcement", &announcement.id.to_string())
-                .details(serde_json::json!({ "action": "create", "title": &announcement.title }))
-                .log();
-            audit.persist(&state_guard.db).await;
-            (
-                StatusCode::CREATED,
-                Json(ApiResponse::success(announcement)),
-            )
-        }
-        Err(e) => {
-            tracing::error!("Failed to save announcement: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to create announcement",
-                )),
-            )
-        }
-    }
-}
-
-/// Represents a field that can be absent, explicitly null, or a value.
-/// This avoids `Option<Option<T>>` which clippy flags.
-#[derive(Default)]
-pub enum NullableField<T> {
-    /// Field was not present in the request
-    #[default]
-    Absent,
-    /// Field was explicitly set to null
-    Null,
-    /// Field was set to a value
-    Value(T),
-}
-
-impl<'de, T: Deserialize<'de>> Deserialize<'de> for NullableField<T> {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        Option::<T>::deserialize(deserializer).map(|opt| opt.map_or(Self::Null, Self::Value))
-    }
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-pub struct UpdateAnnouncementRequest {
-    title: Option<String>,
-    message: Option<String>,
-    severity: Option<AnnouncementSeverity>,
-    active: Option<bool>,
-    #[schema(value_type = Option<String>)]
-    #[serde(default)]
-    expires_at: NullableField<DateTime<Utc>>,
-}
-
-/// `PUT /api/v1/admin/announcements/{id}` — admin: update announcement
-#[utoipa::path(
-    put,
-    path = "/api/v1/admin/announcements/{id}",
-    tag = "Admin",
-    summary = "Update announcement",
-    description = "Update an existing announcement by ID. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_update_announcement(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateAnnouncementRequest>,
-) -> (StatusCode, Json<ApiResponse<Announcement>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    // Fetch all announcements and find by ID
-    let announcements = match state_guard.db.list_announcements().await {
-        Ok(a) => a,
-        Err(e) => {
-            tracing::error!("Failed to list announcements: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    let Some(mut announcement) = announcements.into_iter().find(|a| a.id.to_string() == id) else {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Announcement not found")),
-        );
-    };
-
-    if let Some(title) = req.title {
-        announcement.title = title;
-    }
-    if let Some(message) = req.message {
-        announcement.message = message;
-    }
-    if let Some(severity) = req.severity {
-        announcement.severity = severity;
-    }
-    if let Some(active) = req.active {
-        announcement.active = active;
-    }
-    match req.expires_at {
-        NullableField::Value(v) => announcement.expires_at = Some(v),
-        NullableField::Null => announcement.expires_at = None,
-        NullableField::Absent => {}
-    }
-
-    match state_guard.db.save_announcement(&announcement).await {
-        Ok(()) => (StatusCode::OK, Json(ApiResponse::success(announcement))),
-        Err(e) => {
-            tracing::error!("Failed to update announcement: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to update announcement",
-                )),
-            )
-        }
-    }
-}
-
-/// `DELETE /api/v1/admin/announcements/{id}` — admin: delete announcement
-#[utoipa::path(
-    delete,
-    path = "/api/v1/admin/announcements/{id}",
-    tag = "Admin",
-    summary = "Delete announcement",
-    description = "Delete an announcement by ID. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_delete_announcement(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    match state_guard.db.delete_announcement(&id).await {
-        Ok(true) => (StatusCode::OK, Json(ApiResponse::success(()))),
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Announcement not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to delete announcement: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to delete announcement",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// NOTIFICATIONS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/notifications` — list current user's notifications (most recent 50)
-#[utoipa::path(get, path = "/api/v1/notifications", tag = "Notifications",
-    summary = "List user notifications",
-    description = "Returns recent notifications for the authenticated user.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn list_notifications(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<Notification>>>) {
-    let state_guard = state.read().await;
-    match state_guard
-        .db
-        .list_notifications_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(mut notifications) => {
-            // Sort by created_at descending (most recent first)
-            notifications.sort_by(|a, b| b.created_at.cmp(&a.created_at));
-            notifications.truncate(50);
-            (StatusCode::OK, Json(ApiResponse::success(notifications)))
-        }
-        Err(e) => {
-            tracing::error!("Failed to list notifications: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list notifications",
-                )),
-            )
-        }
-    }
-}
-
-/// `PUT /api/v1/notifications/{id}/read` — mark notification as read (verify ownership)
-#[utoipa::path(put, path = "/api/v1/notifications/{id}/read", tag = "Notifications",
-    summary = "Mark notification as read",
-    description = "Marks a notification as read.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn mark_notification_read(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-
-    // Verify ownership by listing user's notifications
-    let notifications = match state_guard
-        .db
-        .list_notifications_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(n) => n,
-        Err(e) => {
-            tracing::error!("Failed to list notifications: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    let owns = notifications.iter().any(|n| n.id.to_string() == id);
-    if !owns {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Notification not found")),
-        );
-    }
-
-    match state_guard.db.mark_notification_read(&id).await {
-        Ok(true) => (StatusCode::OK, Json(ApiResponse::success(()))),
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Notification not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to mark notification read: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to mark notification read",
-                )),
-            )
-        }
-    }
-}
-
-/// `POST /api/v1/notifications/read-all` — mark all user's notifications as read
-#[utoipa::path(post, path = "/api/v1/notifications/read-all", tag = "Notifications",
-    summary = "Mark all notifications as read",
-    description = "Marks all notifications as read.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn mark_all_notifications_read(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<u32>>) {
-    let state_guard = state.read().await;
-    let notifications = match state_guard
-        .db
-        .list_notifications_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(n) => n,
-        Err(e) => {
-            tracing::error!("Failed to list notifications: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to mark notifications read",
-                )),
-            );
-        }
-    };
-
-    let mut count = 0u32;
-    for notif in &notifications {
-        if !notif.read
-            && matches!(
-                state_guard
-                    .db
-                    .mark_notification_read(&notif.id.to_string())
-                    .await,
-                Ok(true)
-            )
-        {
-            count += 1;
-        }
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(count)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// WAITLIST
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/waitlist` — list current user's waitlist entries
-#[utoipa::path(get, path = "/api/v1/waitlist", tag = "Waitlist",
-    summary = "List waitlist entries",
-    description = "Returns waitlist entries for the authenticated user.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn list_waitlist(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> Json<ApiResponse<Vec<WaitlistEntry>>> {
-    let state_guard = state.read().await;
-    match state_guard
-        .db
-        .list_waitlist_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(entries) => Json(ApiResponse::success(entries)),
-        Err(e) => {
-            tracing::error!("Failed to list waitlist entries: {}", e);
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to list waitlist entries",
-            ))
-        }
-    }
-}
-
-/// Request body for joining the waitlist
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct JoinWaitlistRequest {
-    lot_id: Uuid,
-}
-
-/// `POST /api/v1/waitlist` — join waitlist for a lot
-#[utoipa::path(post, path = "/api/v1/waitlist", tag = "Waitlist",
-    summary = "Join waitlist",
-    description = "Adds the user to a lot waitlist.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn join_waitlist(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<JoinWaitlistRequest>,
-) -> (StatusCode, Json<ApiResponse<WaitlistEntry>>) {
-    let state_guard = state.read().await;
-
-    // Check waitlist_enabled setting
-    let waitlist_enabled = read_admin_setting(&state_guard.db, "waitlist_enabled").await;
-    if waitlist_enabled != "true" {
-        return (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(ApiResponse::error(
-                "WAITLIST_DISABLED",
-                "Waitlist is not enabled",
-            )),
-        );
-    }
-
-    // First-or-create: check if user already has a waitlist entry for this lot
-    let existing = state_guard
-        .db
-        .list_waitlist_by_user(&auth_user.user_id.to_string())
-        .await
-        .unwrap_or_default();
-    if let Some(entry) = existing.iter().find(|e| e.lot_id == req.lot_id) {
-        return (StatusCode::OK, Json(ApiResponse::success(entry.clone())));
-    }
-
-    let entry = WaitlistEntry {
-        id: Uuid::new_v4(),
-        user_id: auth_user.user_id,
-        lot_id: req.lot_id,
-        created_at: Utc::now(),
-        notified_at: None,
-    };
-
-    if let Err(e) = state_guard.db.save_waitlist_entry(&entry).await {
-        tracing::error!("Failed to save waitlist entry: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to join waitlist",
-            )),
-        );
-    }
-
-    (StatusCode::CREATED, Json(ApiResponse::success(entry)))
-}
-
-/// `DELETE /api/v1/waitlist/{id}` — leave waitlist (verify ownership)
-#[utoipa::path(delete, path = "/api/v1/waitlist/{id}", tag = "Waitlist",
-    summary = "Leave waitlist",
-    description = "Removes the user from a waitlist entry.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn leave_waitlist(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-
-    // Verify ownership
-    match state_guard.db.get_waitlist_entry(&id).await {
-        Ok(Some(entry)) => {
-            if entry.user_id != auth_user.user_id {
-                return (
-                    StatusCode::FORBIDDEN,
-                    Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-                );
-            }
-        }
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Waitlist entry not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    }
-
-    match state_guard.db.delete_waitlist_entry(&id).await {
-        Ok(true) => (StatusCode::OK, Json(ApiResponse::success(()))),
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error("NOT_FOUND", "Waitlist entry not found")),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to delete waitlist entry: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to leave waitlist",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// SWAP REQUESTS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/swap-requests` — list user's swap requests (as requester or target)
-#[utoipa::path(
-    get,
-    path = "/api/v1/swap-requests",
-    tag = "Bookings",
-    summary = "List swap requests",
-    description = "List the current user's swap requests (as requester or target).",
-    security(("bearer_auth" = []))
-)]
-pub async fn list_swap_requests(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> Json<ApiResponse<Vec<SwapRequest>>> {
-    let state_guard = state.read().await;
-    match state_guard
-        .db
-        .list_swap_requests_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(requests) => Json(ApiResponse::success(requests)),
-        Err(e) => {
-            tracing::error!("Failed to list swap requests: {}", e);
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to list swap requests",
-            ))
-        }
-    }
-}
-
-/// Request body for creating a swap request
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct CreateSwapRequestBody {
-    target_booking_id: Uuid,
-    message: Option<String>,
-}
-
-/// `POST /api/v1/bookings/{id}/swap-request` — create a swap request
-#[utoipa::path(
-    post,
-    path = "/api/v1/bookings/{id}/swap-request",
-    tag = "Bookings",
-    summary = "Create swap request",
-    description = "Create a parking slot swap request for a booking.",
-    security(("bearer_auth" = []))
-)]
-pub async fn create_swap_request(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(booking_id): Path<String>,
-    Json(req): Json<CreateSwapRequestBody>,
-) -> (StatusCode, Json<ApiResponse<SwapRequest>>) {
-    let state_guard = state.read().await;
-
-    // Get requester's booking
-    let requester_booking = match state_guard.db.get_booking(&booking_id).await {
-        Ok(Some(b)) => b,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Booking not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Verify ownership of requester booking
-    if requester_booking.user_id != auth_user.user_id {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error(
-                "FORBIDDEN",
-                "You can only create swap requests for your own bookings",
-            )),
-        );
-    }
-
-    // Get target booking
-    let target_booking = match state_guard
-        .db
-        .get_booking(&req.target_booking_id.to_string())
-        .await
-    {
-        Ok(Some(b)) => b,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Target booking not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Validate: different users
-    if requester_booking.user_id == target_booking.user_id {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_SWAP",
-                "Cannot swap with your own booking",
-            )),
-        );
-    }
-
-    // Validate: same lot
-    if requester_booking.lot_id != target_booking.lot_id {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_SWAP",
-                "Bookings must be in the same lot",
-            )),
-        );
-    }
-
-    let swap_request = SwapRequest {
-        id: Uuid::new_v4(),
-        requester_booking_id: requester_booking.id,
-        target_booking_id: target_booking.id,
-        requester_id: auth_user.user_id,
-        target_id: target_booking.user_id,
-        status: SwapRequestStatus::Pending,
-        message: req.message,
-        created_at: Utc::now(),
-    };
-
-    if let Err(e) = state_guard.db.save_swap_request(&swap_request).await {
-        tracing::error!("Failed to save swap request: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to create swap request",
-            )),
-        );
-    }
-
-    (
-        StatusCode::CREATED,
-        Json(ApiResponse::success(swap_request)),
-    )
-}
-
-/// Request body for accepting/declining a swap request
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateSwapRequestBody {
-    action: String,
-}
-
-/// `PUT /api/v1/swap-requests/{id}` — accept or decline a swap request
-#[utoipa::path(
-    put,
-    path = "/api/v1/swap-requests/{id}",
-    tag = "Bookings",
-    summary = "Update swap request",
-    description = "Accept or decline a swap request.",
-    security(("bearer_auth" = []))
-)]
-pub async fn update_swap_request(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateSwapRequestBody>,
-) -> (StatusCode, Json<ApiResponse<SwapRequest>>) {
-    // Use write lock for atomic swap if accepting
-    let state_guard = state.write().await;
-
-    let mut swap = match state_guard.db.get_swap_request(&id).await {
-        Ok(Some(s)) => s,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Swap request not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Only the target user can accept/decline
-    if swap.target_id != auth_user.user_id {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error(
-                "FORBIDDEN",
-                "Only the target user can respond to this swap request",
-            )),
-        );
-    }
-
-    if swap.status != SwapRequestStatus::Pending {
-        return (
-            StatusCode::CONFLICT,
-            Json(ApiResponse::error(
-                "ALREADY_RESOLVED",
-                "This swap request has already been resolved",
-            )),
-        );
-    }
-
-    match req.action.as_str() {
-        "accept" => {
-            // Get both bookings
-            let Ok(Some(mut requester_booking)) = state_guard
-                .db
-                .get_booking(&swap.requester_booking_id.to_string())
-                .await
-            else {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ApiResponse::error(
-                        "SERVER_ERROR",
-                        "Requester booking not found",
-                    )),
-                );
-            };
-
-            let Ok(Some(mut target_booking)) = state_guard
-                .db
-                .get_booking(&swap.target_booking_id.to_string())
-                .await
-            else {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ApiResponse::error(
-                        "SERVER_ERROR",
-                        "Target booking not found",
-                    )),
-                );
-            };
-
-            // Swap slot_ids between the two bookings
-            std::mem::swap(&mut requester_booking.slot_id, &mut target_booking.slot_id);
-            std::mem::swap(
-                &mut requester_booking.slot_number,
-                &mut target_booking.slot_number,
-            );
-            std::mem::swap(
-                &mut requester_booking.floor_name,
-                &mut target_booking.floor_name,
-            );
-            let now = Utc::now();
-            requester_booking.updated_at = now;
-            target_booking.updated_at = now;
-
-            if let Err(e) = state_guard.db.save_booking(&requester_booking).await {
-                tracing::error!("Failed to save requester booking during swap: {}", e);
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ApiResponse::error("SERVER_ERROR", "Failed to perform swap")),
-                );
-            }
-            if let Err(e) = state_guard.db.save_booking(&target_booking).await {
-                tracing::error!("Failed to save target booking during swap: {}", e);
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ApiResponse::error("SERVER_ERROR", "Failed to perform swap")),
-                );
-            }
-
-            swap.status = SwapRequestStatus::Accepted;
-        }
-        "decline" => {
-            swap.status = SwapRequestStatus::Declined;
-        }
-        _ => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_ACTION",
-                    "Action must be 'accept' or 'decline'",
-                )),
-            );
-        }
-    }
-
-    if let Err(e) = state_guard.db.save_swap_request(&swap).await {
-        tracing::error!("Failed to update swap request: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to update swap request",
-            )),
-        );
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(swap)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// RECURRING BOOKINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/recurring-bookings` — list user's recurring bookings
-#[utoipa::path(
-    get,
-    path = "/api/v1/recurring-bookings",
-    tag = "Bookings",
-    summary = "List recurring bookings",
-    description = "List the current user's recurring booking patterns.",
-    security(("bearer_auth" = []))
-)]
-pub async fn list_recurring_bookings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> Json<ApiResponse<Vec<RecurringBooking>>> {
-    let state_guard = state.read().await;
-    match state_guard
-        .db
-        .list_recurring_bookings_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(bookings) => Json(ApiResponse::success(bookings)),
-        Err(e) => {
-            tracing::error!("Failed to list recurring bookings: {}", e);
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to list recurring bookings",
-            ))
-        }
-    }
-}
-
-/// Request body for creating a recurring booking
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct CreateRecurringBookingRequest {
-    lot_id: Uuid,
-    slot_id: Option<Uuid>,
-    days_of_week: Vec<u8>,
-    start_date: String,
-    end_date: Option<String>,
-    start_time: String,
-    end_time: String,
-    vehicle_plate: Option<String>,
-}
-
-/// `POST /api/v1/recurring-bookings` — create a recurring booking
-#[utoipa::path(
-    post,
-    path = "/api/v1/recurring-bookings",
-    tag = "Bookings",
-    summary = "Create recurring booking",
-    description = "Create a new recurring booking pattern (e.g. every Tuesday 8-17).",
-    security(("bearer_auth" = []))
-)]
-pub async fn create_recurring_booking(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<CreateRecurringBookingRequest>,
-) -> (StatusCode, Json<ApiResponse<RecurringBooking>>) {
-    let state_guard = state.read().await;
-
-    let booking = RecurringBooking {
-        id: Uuid::new_v4(),
-        user_id: auth_user.user_id,
-        lot_id: req.lot_id,
-        slot_id: req.slot_id,
-        days_of_week: req.days_of_week,
-        start_date: req.start_date,
-        end_date: req.end_date,
-        start_time: req.start_time,
-        end_time: req.end_time,
-        vehicle_plate: req.vehicle_plate,
-        active: true,
-        created_at: Utc::now(),
-    };
-
-    if let Err(e) = state_guard.db.save_recurring_booking(&booking).await {
-        tracing::error!("Failed to save recurring booking: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to create recurring booking",
-            )),
-        );
-    }
-
-    (StatusCode::CREATED, Json(ApiResponse::success(booking)))
-}
-
-/// `DELETE /api/v1/recurring-bookings/{id}` — delete recurring booking (verify ownership)
-#[utoipa::path(
-    delete,
-    path = "/api/v1/recurring-bookings/{id}",
-    tag = "Bookings",
-    summary = "Delete recurring booking",
-    description = "Delete a recurring booking pattern. Verifies ownership.",
-    security(("bearer_auth" = []))
-)]
-pub async fn delete_recurring_booking(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<()>>) {
-    let state_guard = state.read().await;
-
-    // Check ownership via listing user's recurring bookings
-    let user_bookings = state_guard
-        .db
-        .list_recurring_bookings_by_user(&auth_user.user_id.to_string())
-        .await
-        .unwrap_or_default();
-
-    let Ok(id_uuid) = Uuid::parse_str(&id) else {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error("INVALID_ID", "Invalid ID format")),
-        );
-    };
-
-    if !user_bookings.iter().any(|b| b.id == id_uuid) {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    }
-
-    match state_guard.db.delete_recurring_booking(&id).await {
-        Ok(true) => (StatusCode::OK, Json(ApiResponse::success(()))),
-        Ok(false) => (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error(
-                "NOT_FOUND",
-                "Recurring booking not found",
-            )),
-        ),
-        Err(e) => {
-            tracing::error!("Failed to delete recurring booking: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to delete recurring booking",
-                )),
-            )
-        }
-    }
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// GUEST BOOKINGS
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for creating a guest booking
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct CreateGuestBookingRequest {
-    lot_id: Uuid,
-    slot_id: Uuid,
-    start_time: chrono::DateTime<Utc>,
-    end_time: chrono::DateTime<Utc>,
-    guest_name: String,
-    guest_email: Option<String>,
-}
-
-/// Generate an 8-character random alphanumeric guest code
-fn generate_guest_code() -> String {
-    use rand::Rng;
-    const CHARSET: &[u8] = b"ABCDEFGHJKLMNPQRSTUVWXYZ23456789";
-    let mut rng = rand::rng();
-    (0..8)
-        .map(|_| {
-            let idx = rng.random_range(0..CHARSET.len());
-            CHARSET[idx] as char
-        })
-        .collect()
-}
-
-/// `POST /api/v1/bookings/guest` — create a guest booking
-#[utoipa::path(
-    post,
-    path = "/api/v1/bookings/guest",
-    tag = "Bookings",
-    summary = "Create guest booking",
-    description = "Create a visitor parking booking with a guest code.",
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state, req), fields(user_id = %auth_user.user_id, guest_name = %req.guest_name))]
-pub async fn create_guest_booking(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Json(req): Json<CreateGuestBookingRequest>,
-) -> (StatusCode, Json<ApiResponse<GuestBooking>>) {
-    let state_guard = state.read().await;
-
-    // Check allow_guest_bookings setting
-    let allowed = read_admin_setting(&state_guard.db, "allow_guest_bookings").await;
-    if allowed != "true" {
-        return (
-            StatusCode::UNPROCESSABLE_ENTITY,
-            Json(ApiResponse::error(
-                "GUEST_BOOKINGS_DISABLED",
-                "Guest bookings are not enabled",
-            )),
-        );
-    }
-
-    let guest_booking = GuestBooking {
-        id: Uuid::new_v4(),
-        created_by: auth_user.user_id,
-        lot_id: req.lot_id,
-        slot_id: req.slot_id,
-        guest_name: req.guest_name,
-        guest_email: req.guest_email,
-        guest_code: generate_guest_code(),
-        start_time: req.start_time,
-        end_time: req.end_time,
-        vehicle_plate: None,
-        status: BookingStatus::Confirmed,
-        created_at: Utc::now(),
-    };
-
-    if let Err(e) = state_guard.db.save_guest_booking(&guest_booking).await {
-        tracing::error!("Failed to save guest booking: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to create guest booking",
-            )),
-        );
-    }
-
-    (
-        StatusCode::CREATED,
-        Json(ApiResponse::success(guest_booking)),
-    )
-}
-
-/// `GET /api/v1/admin/guest-bookings` — admin: list all guest bookings
-#[utoipa::path(
-    get,
-    path = "/api/v1/admin/guest-bookings",
-    tag = "Admin",
-    summary = "List guest bookings",
-    description = "List all guest bookings. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_list_guest_bookings(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> (StatusCode, Json<ApiResponse<Vec<GuestBooking>>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    match state_guard.db.list_guest_bookings().await {
-        Ok(bookings) => (StatusCode::OK, Json(ApiResponse::success(bookings))),
-        Err(e) => {
-            tracing::error!("Failed to list guest bookings: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to list guest bookings",
-                )),
-            )
-        }
-    }
-}
-
-/// `PATCH /api/v1/admin/guest-bookings/{id}/cancel` — admin: cancel a guest booking
-#[utoipa::path(
-    patch,
-    path = "/api/v1/admin/guest-bookings/{id}/cancel",
-    tag = "Admin",
-    summary = "Cancel guest booking",
-    description = "Cancel a guest booking by ID. Admin only.",
-    security(("bearer_auth" = []))
-)]
-pub async fn admin_cancel_guest_booking(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-) -> (StatusCode, Json<ApiResponse<GuestBooking>>) {
-    let state_guard = state.read().await;
-    if let Err((status, msg)) = check_admin(&state_guard, &auth_user).await {
-        return (status, Json(ApiResponse::error("FORBIDDEN", msg)));
-    }
-
-    let mut booking = match state_guard.db.get_guest_booking(&id).await {
-        Ok(Some(b)) => b,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Guest booking not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    booking.status = BookingStatus::Cancelled;
-
-    if let Err(e) = state_guard.db.save_guest_booking(&booking).await {
-        tracing::error!("Failed to cancel guest booking: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to cancel guest booking",
-            )),
-        );
-    }
-
-    (StatusCode::OK, Json(ApiResponse::success(booking)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
 // QUICK BOOK
 // ═══════════════════════════════════════════════════════════════════════════════
 
@@ -5654,141 +3590,6 @@ pub async fn quick_book(
     );
 
     (StatusCode::CREATED, Json(ApiResponse::success(booking)))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// CALENDAR
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Query params for calendar events
-#[derive(Debug, Deserialize)]
-pub struct CalendarQuery {
-    from: Option<String>,
-    to: Option<String>,
-}
-
-/// Calendar event response
-#[derive(Debug, Serialize)]
-pub struct CalendarEvent {
-    id: String,
-    #[serde(rename = "type")]
-    event_type: String,
-    title: String,
-    start: chrono::DateTime<Utc>,
-    end: chrono::DateTime<Utc>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    lot_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    slot_number: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    status: Option<String>,
-}
-
-/// `GET /api/v1/calendar/events` — return user's bookings + absences as calendar events
-#[utoipa::path(get, path = "/api/v1/calendar/events", tag = "Calendar",
-    summary = "Calendar events",
-    description = "Returns bookings and absences as calendar events.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn calendar_events(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Query(query): Query<CalendarQuery>,
-) -> Json<ApiResponse<Vec<CalendarEvent>>> {
-    let state_guard = state.read().await;
-    let mut events = Vec::new();
-
-    // Parse date range for filtering
-    let from_date = query
-        .from
-        .as_deref()
-        .and_then(|s| chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").ok());
-    let to_date = query
-        .to
-        .as_deref()
-        .and_then(|s| chrono::NaiveDate::parse_from_str(s, "%Y-%m-%d").ok());
-
-    // Bookings as events
-    if let Ok(bookings) = state_guard
-        .db
-        .list_bookings_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        for b in bookings {
-            // Filter by date range if provided
-            if let Some(from) = from_date {
-                if b.start_time.date_naive() < from {
-                    continue;
-                }
-            }
-            if let Some(to) = to_date {
-                if b.start_time.date_naive() > to {
-                    continue;
-                }
-            }
-
-            events.push(CalendarEvent {
-                id: b.id.to_string(),
-                event_type: "booking".to_string(),
-                title: format!("Parking - Slot {}", b.slot_number),
-                start: b.start_time,
-                end: b.end_time,
-                lot_name: Some(b.floor_name.clone()),
-                slot_number: Some(b.slot_number),
-                status: Some(format!("{:?}", b.status).to_lowercase()),
-            });
-        }
-    }
-
-    // Absences as events
-    if let Ok(absences) = state_guard
-        .db
-        .list_absences_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        for a in absences {
-            let start = chrono::NaiveDate::parse_from_str(&a.start_date, "%Y-%m-%d")
-                .ok()
-                .and_then(|d| d.and_hms_opt(0, 0, 0))
-                .map(|dt| chrono::DateTime::<Utc>::from_naive_utc_and_offset(dt, Utc));
-            let end = chrono::NaiveDate::parse_from_str(&a.end_date, "%Y-%m-%d")
-                .ok()
-                .and_then(|d| d.and_hms_opt(23, 59, 59))
-                .map(|dt| chrono::DateTime::<Utc>::from_naive_utc_and_offset(dt, Utc));
-
-            if let (Some(start_dt), Some(end_dt)) = (start, end) {
-                // Filter by date range
-                if let Some(from) = from_date {
-                    if end_dt.date_naive() < from {
-                        continue;
-                    }
-                }
-                if let Some(to) = to_date {
-                    if start_dt.date_naive() > to {
-                        continue;
-                    }
-                }
-
-                let type_label = format!("{:?}", a.absence_type);
-                events.push(CalendarEvent {
-                    id: a.id.to_string(),
-                    event_type: "absence".to_string(),
-                    title: type_label,
-                    start: start_dt,
-                    end: end_dt,
-                    lot_name: None,
-                    slot_number: None,
-                    status: None,
-                });
-            }
-        }
-    }
-
-    // Sort by start time
-    events.sort_by(|a, b| a.start.cmp(&b.start));
-
-    Json(ApiResponse::success(events))
 }
 
 // ═══════════════════════════════════════════════════════════════════════════════
@@ -6020,48 +3821,6 @@ pub async fn admin_audit_log(
 }
 
 // ═══════════════════════════════════════════════════════════════════════════════
-// TEAM LIST
-// ═══════════════════════════════════════════════════════════════════════════════
-
-#[derive(Serialize)]
-pub struct TeamMember {
-    id: Uuid,
-    name: String,
-    username: String,
-    role: String,
-    is_active: bool,
-}
-
-/// `GET /api/v1/team` — list all team members (simplified view)
-#[utoipa::path(get, path = "/api/v1/team", tag = "Team",
-    summary = "List team members",
-    description = "Returns all active team members.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn team_list(
-    State(state): State<SharedState>,
-    Extension(_auth_user): Extension<AuthUser>,
-) -> Json<ApiResponse<Vec<TeamMember>>> {
-    let state_guard = state.read().await;
-
-    let users = state_guard.db.list_users().await.unwrap_or_default();
-    let members: Vec<TeamMember> = users
-        .into_iter()
-        .filter(|u| u.is_active)
-        .map(|u| TeamMember {
-            id: u.id,
-            name: u.name,
-            username: u.username,
-            role: format!("{:?}", u.role).to_lowercase(),
-            is_active: u.is_active,
-        })
-        .collect();
-
-    Json(ApiResponse::success(members))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
 // CHANGE PASSWORD
 // ═══════════════════════════════════════════════════════════════════════════════
 
@@ -6158,72 +3917,6 @@ pub async fn change_password(
     }
 
     (StatusCode::OK, Json(ApiResponse::success(())))
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// iCAL EXPORT
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// `GET /api/v1/user/calendar.ics` — export user's bookings as iCal
-#[utoipa::path(get, path = "/api/v1/user/calendar.ics", tag = "Calendar",
-    summary = "Export bookings as iCal",
-    description = "Returns bookings in iCalendar format.",
-    security(("bearer_auth" = [])),
-    responses((status = 200, description = "Success"))
-)]
-pub async fn user_calendar_ics(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-) -> impl axum::response::IntoResponse {
-    let state_guard = state.read().await;
-
-    let bookings = match state_guard
-        .db
-        .list_bookings_by_user(&auth_user.user_id.to_string())
-        .await
-    {
-        Ok(b) => b,
-        Err(e) => {
-            tracing::error!("Failed to list bookings for iCal: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                [(header::CONTENT_TYPE, "text/plain; charset=utf-8")],
-                "Failed to generate calendar".to_string(),
-            );
-        }
-    };
-
-    let mut ical = String::from("BEGIN:VCALENDAR\r\nVERSION:2.0\r\nPRODID:-//ParkHub//EN\r\n");
-
-    for b in &bookings {
-        // Resolve lot name (best-effort)
-        let lot_name = match state_guard.db.get_parking_lot(&b.lot_id.to_string()).await {
-            Ok(Some(l)) => l.name,
-            _ => "Unknown Lot".to_string(),
-        };
-
-        let dtstart = b.start_time.format("%Y%m%dT%H%M%SZ");
-        let dtend = b.end_time.format("%Y%m%dT%H%M%SZ");
-
-        ical.push_str("BEGIN:VEVENT\r\n");
-        let _ = write!(ical, "UID:{}@parkhub\r\n", b.id);
-        let _ = write!(ical, "DTSTART:{dtstart}\r\n");
-        let _ = write!(ical, "DTEND:{dtend}\r\n");
-        let _ = write!(
-            ical,
-            "SUMMARY:Parking - {} - Slot {}\r\n",
-            lot_name, b.slot_number
-        );
-        ical.push_str("END:VEVENT\r\n");
-    }
-
-    ical.push_str("END:VCALENDAR\r\n");
-
-    (
-        StatusCode::OK,
-        [(header::CONTENT_TYPE, "text/calendar; charset=utf-8")],
-        ical,
-    )
 }
 
 // ═══════════════════════════════════════════════════════════════════════════════
@@ -7434,234 +5127,6 @@ pub async fn update_booking(
 
     (StatusCode::OK, Json(ApiResponse::success(booking)))
 }
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// RECURRING BOOKING UPDATE (PUT /api/v1/recurring-bookings/{id})
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for updating a recurring booking
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateRecurringBookingRequest {
-    pub days_of_week: Option<Vec<u8>>,
-    pub start_date: Option<String>,
-    pub end_date: Option<String>,
-}
-
-/// `PUT /api/v1/recurring-bookings/{id}` — update a recurring booking pattern
-#[utoipa::path(
-    put,
-    path = "/api/v1/recurring-bookings/{id}",
-    tag = "Bookings",
-    summary = "Update a recurring booking",
-    description = "Update days_of_week, start_date, or end_date of a recurring booking. Only the owner or an admin may update. Note: re-expansion of future bookings is not performed automatically.",
-    security(("bearer_auth" = []))
-)]
-pub async fn update_recurring_booking(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateRecurringBookingRequest>,
-) -> (StatusCode, Json<ApiResponse<serde_json::Value>>) {
-    let Ok(id_uuid) = Uuid::parse_str(&id) else {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error("INVALID_ID", "Invalid ID format")),
-        );
-    };
-
-    let state_guard = state.read().await;
-
-    // Fetch caller to check admin status
-    let Ok(Some(caller)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    };
-    let is_admin = caller.role == UserRole::Admin || caller.role == UserRole::SuperAdmin;
-
-    // Try ownership lookup first
-    let user_bookings = state_guard
-        .db
-        .list_recurring_bookings_by_user(&auth_user.user_id.to_string())
-        .await
-        .unwrap_or_default();
-
-    let Some(mut booking) = user_bookings.into_iter().find(|b| b.id == id_uuid) else {
-        if !is_admin {
-            return (
-                StatusCode::FORBIDDEN,
-                Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-            );
-        }
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ApiResponse::error(
-                "NOT_FOUND",
-                "Recurring booking not found",
-            )),
-        );
-    };
-
-    if let Some(days) = req.days_of_week {
-        booking.days_of_week = days;
-    }
-    if let Some(start_date) = req.start_date {
-        booking.start_date = start_date;
-    }
-    if let Some(end_date) = req.end_date {
-        booking.end_date = Some(end_date);
-    }
-
-    if let Err(e) = state_guard.db.save_recurring_booking(&booking).await {
-        tracing::error!("Failed to update recurring booking: {}", e);
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ApiResponse::error(
-                "SERVER_ERROR",
-                "Failed to update recurring booking",
-            )),
-        );
-    }
-
-    (
-        StatusCode::OK,
-        Json(ApiResponse::success(serde_json::json!({
-            "id": booking.id.to_string(),
-            "days_of_week": booking.days_of_week,
-            "start_date": booking.start_date,
-            "end_date": booking.end_date,
-            "note": "Future expanded bookings are not re-generated automatically. Trigger re-expansion separately if needed."
-        }))),
-    )
-}
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// ABSENCE UPDATE (PUT /api/v1/absences/{id})
-// ═══════════════════════════════════════════════════════════════════════════════
-
-/// Request body for updating an absence
-#[derive(Debug, Deserialize, utoipa::ToSchema)]
-pub struct UpdateAbsenceRequest {
-    pub absence_type: Option<AbsenceType>,
-    pub start_date: Option<String>,
-    pub end_date: Option<String>,
-    pub notes: Option<String>,
-}
-
-/// `PUT /api/v1/absences/{id}` — update an existing absence
-#[utoipa::path(
-    put,
-    path = "/api/v1/absences/{id}",
-    tag = "Absences",
-    summary = "Update an absence",
-    description = "Update absence_type, start_date, end_date, or notes. Only the owner or an admin may update.",
-    security(("bearer_auth" = []))
-)]
-pub async fn update_absence(
-    State(state): State<SharedState>,
-    Extension(auth_user): Extension<AuthUser>,
-    Path(id): Path<String>,
-    Json(req): Json<UpdateAbsenceRequest>,
-) -> (StatusCode, Json<ApiResponse<Absence>>) {
-    let state_guard = state.read().await;
-
-    let mut absence = match state_guard.db.get_absence(&id).await {
-        Ok(Some(a)) => a,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ApiResponse::error("NOT_FOUND", "Absence not found")),
-            );
-        }
-        Err(e) => {
-            tracing::error!("Database error fetching absence: {}", e);
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error("SERVER_ERROR", "Internal server error")),
-            );
-        }
-    };
-
-    // Check ownership or admin
-    let Ok(Some(caller)) = state_guard
-        .db
-        .get_user(&auth_user.user_id.to_string())
-        .await
-    else {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    };
-    let is_admin = caller.role == UserRole::Admin || caller.role == UserRole::SuperAdmin;
-    if absence.user_id != auth_user.user_id && !is_admin {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(ApiResponse::error("FORBIDDEN", "Access denied")),
-        );
-    }
-
-    if let Some(absence_type) = req.absence_type {
-        absence.absence_type = absence_type;
-    }
-    if let Some(start_date) = req.start_date {
-        if !is_valid_date(&start_date) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "start_date must be in YYYY-MM-DD format",
-                )),
-            );
-        }
-        absence.start_date = start_date;
-    }
-    if let Some(end_date) = req.end_date {
-        if !is_valid_date(&end_date) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ApiResponse::error(
-                    "INVALID_INPUT",
-                    "end_date must be in YYYY-MM-DD format",
-                )),
-            );
-        }
-        absence.end_date = end_date;
-    }
-    if let Some(notes) = req.notes {
-        absence.note = Some(notes);
-    }
-
-    if absence.start_date > absence.end_date {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ApiResponse::error(
-                "INVALID_INPUT",
-                "start_date must not be after end_date",
-            )),
-        );
-    }
-
-    match state_guard.db.save_absence(&absence).await {
-        Ok(()) => (StatusCode::OK, Json(ApiResponse::success(absence))),
-        Err(e) => {
-            tracing::error!("Failed to update absence: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ApiResponse::error(
-                    "SERVER_ERROR",
-                    "Failed to update absence",
-                )),
-            )
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -7696,62 +5161,6 @@ mod tests {
     fn test_normalize_metric_path_root() {
         assert_eq!(normalize_metric_path("/health"), "/health");
         assert_eq!(normalize_metric_path("/metrics"), "/metrics");
-    }
-
-    // ─── is_valid_date ─────────────────────────────────────────────────
-
-    #[test]
-    fn test_is_valid_date_correct() {
-        assert!(is_valid_date("2026-03-20"));
-        assert!(is_valid_date("2000-01-01"));
-        assert!(is_valid_date("2026-12-31"));
-    }
-
-    #[test]
-    fn test_is_valid_date_invalid() {
-        assert!(!is_valid_date("2026-13-01")); // month 13
-        assert!(!is_valid_date("2026-02-30")); // Feb 30
-        assert!(!is_valid_date("not-a-date"));
-        assert!(!is_valid_date(""));
-        assert!(!is_valid_date("20260320"));
-        assert!(!is_valid_date("2026/03/20"));
-    }
-
-    #[test]
-    fn test_is_valid_date_leap_year() {
-        assert!(is_valid_date("2024-02-29")); // 2024 is leap
-        assert!(!is_valid_date("2025-02-29")); // 2025 is not
-    }
-
-    // ─── generate_guest_code ───────────────────────────────────────────
-
-    #[test]
-    fn test_generate_guest_code_length() {
-        let code = generate_guest_code();
-        assert_eq!(code.len(), 8);
-    }
-
-    #[test]
-    fn test_generate_guest_code_charset() {
-        let valid_chars: &str = "ABCDEFGHJKLMNPQRSTUVWXYZ23456789";
-        for _ in 0..20 {
-            let code = generate_guest_code();
-            for c in code.chars() {
-                assert!(
-                    valid_chars.contains(c),
-                    "Invalid char '{}' in guest code",
-                    c
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_generate_guest_code_uniqueness() {
-        let codes: Vec<String> = (0..50).map(|_| generate_guest_code()).collect();
-        let unique: std::collections::HashSet<&String> = codes.iter().collect();
-        // With 8 chars from 31-char set, collisions in 50 codes are astronomically unlikely
-        assert!(unique.len() > 45);
     }
 
     // ─── ImpressumData serde ───────────────────────────────────────────
@@ -7830,120 +5239,6 @@ mod tests {
         let req: ChangePasswordRequest = serde_json::from_str(json).unwrap();
         assert_eq!(req.current_password, "old");
         assert_eq!(req.new_password, "NewSecure123!");
-    }
-
-    // ─── JoinWaitlistRequest ───────────────────────────────────────────
-
-    #[test]
-    fn test_join_waitlist_request() {
-        let json = r#"{"lot_id":"550e8400-e29b-41d4-a716-446655440000"}"#;
-        let req: JoinWaitlistRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(
-            req.lot_id.to_string(),
-            "550e8400-e29b-41d4-a716-446655440000"
-        );
-    }
-
-    // ─── CreateSwapRequestBody ─────────────────────────────────────────
-
-    #[test]
-    fn test_create_swap_request_body() {
-        let json = r#"{"target_booking_id":"550e8400-e29b-41d4-a716-446655440000","message":"Please swap?"}"#;
-        let req: CreateSwapRequestBody = serde_json::from_str(json).unwrap();
-        assert_eq!(
-            req.target_booking_id.to_string(),
-            "550e8400-e29b-41d4-a716-446655440000"
-        );
-        assert_eq!(req.message.as_deref(), Some("Please swap?"));
-    }
-
-    #[test]
-    fn test_create_swap_request_body_no_message() {
-        let json = r#"{"target_booking_id":"550e8400-e29b-41d4-a716-446655440000"}"#;
-        let req: CreateSwapRequestBody = serde_json::from_str(json).unwrap();
-        assert!(req.message.is_none());
-    }
-
-    // ─── UpdateSwapRequestBody ─────────────────────────────────────────
-
-    #[test]
-    fn test_update_swap_request_body_accept() {
-        let json = r#"{"action":"accept"}"#;
-        let req: UpdateSwapRequestBody = serde_json::from_str(json).unwrap();
-        assert_eq!(req.action, "accept");
-    }
-
-    #[test]
-    fn test_update_swap_request_body_decline() {
-        let json = r#"{"action":"decline"}"#;
-        let req: UpdateSwapRequestBody = serde_json::from_str(json).unwrap();
-        assert_eq!(req.action, "decline");
-    }
-
-    // ─── CreateRecurringBookingRequest ─────────────────────────────────
-
-    #[test]
-    fn test_create_recurring_booking_request_full() {
-        let json = r#"{
-            "lot_id":"550e8400-e29b-41d4-a716-446655440000",
-            "slot_id":"660e8400-e29b-41d4-a716-446655440001",
-            "days_of_week":[1,3,5],
-            "start_date":"2026-04-01",
-            "end_date":"2026-06-30",
-            "start_time":"08:00",
-            "end_time":"17:00",
-            "vehicle_plate":"B-AB 1234"
-        }"#;
-        let req: CreateRecurringBookingRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.days_of_week, vec![1, 3, 5]);
-        assert_eq!(req.start_date, "2026-04-01");
-        assert_eq!(req.end_date.as_deref(), Some("2026-06-30"));
-        assert_eq!(req.vehicle_plate.as_deref(), Some("B-AB 1234"));
-    }
-
-    #[test]
-    fn test_create_recurring_booking_request_minimal() {
-        let json = r#"{
-            "lot_id":"550e8400-e29b-41d4-a716-446655440000",
-            "days_of_week":[1],
-            "start_date":"2026-04-01",
-            "start_time":"09:00",
-            "end_time":"18:00"
-        }"#;
-        let req: CreateRecurringBookingRequest = serde_json::from_str(json).unwrap();
-        assert!(req.slot_id.is_none());
-        assert!(req.end_date.is_none());
-        assert!(req.vehicle_plate.is_none());
-    }
-
-    // ─── CreateGuestBookingRequest ─────────────────────────────────────
-
-    #[test]
-    fn test_create_guest_booking_request() {
-        let json = r#"{
-            "lot_id":"550e8400-e29b-41d4-a716-446655440000",
-            "slot_id":"660e8400-e29b-41d4-a716-446655440001",
-            "start_time":"2026-04-01T08:00:00Z",
-            "end_time":"2026-04-01T17:00:00Z",
-            "guest_name":"Visitor One",
-            "guest_email":"visitor@example.com"
-        }"#;
-        let req: CreateGuestBookingRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.guest_name, "Visitor One");
-        assert_eq!(req.guest_email.as_deref(), Some("visitor@example.com"));
-    }
-
-    #[test]
-    fn test_create_guest_booking_request_no_email() {
-        let json = r#"{
-            "lot_id":"550e8400-e29b-41d4-a716-446655440000",
-            "slot_id":"660e8400-e29b-41d4-a716-446655440001",
-            "start_time":"2026-04-01T08:00:00Z",
-            "end_time":"2026-04-01T17:00:00Z",
-            "guest_name":"Walk-in"
-        }"#;
-        let req: CreateGuestBookingRequest = serde_json::from_str(json).unwrap();
-        assert!(req.guest_email.is_none());
     }
 
     // ─── AdminResetRequest ─────────────────────────────────────────────
@@ -8043,68 +5338,6 @@ mod tests {
         assert_eq!(req.is_active, Some(true));
     }
 
-    // ─── UpdateFeaturesRequest ─────────────────────────────────────────
-
-    #[test]
-    fn test_update_features_request() {
-        let json = r#"{"enabled":["credits","absences","vehicles"]}"#;
-        let req: UpdateFeaturesRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.enabled.len(), 3);
-        assert!(req.enabled.contains(&"credits".to_string()));
-    }
-
-    #[test]
-    fn test_update_features_request_empty() {
-        let json = r#"{"enabled":[]}"#;
-        let req: UpdateFeaturesRequest = serde_json::from_str(json).unwrap();
-        assert!(req.enabled.is_empty());
-    }
-
-    // ─── CreateAbsenceRequest ──────────────────────────────────────────
-
-    #[test]
-    fn test_create_absence_request() {
-        let json = r#"{"absence_type":"homeoffice","start_date":"2026-04-01","end_date":"2026-04-01","note":"WFH"}"#;
-        let req: CreateAbsenceRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.start_date, "2026-04-01");
-        assert_eq!(req.end_date, "2026-04-01");
-        assert_eq!(req.note.as_deref(), Some("WFH"));
-    }
-
-    #[test]
-    fn test_create_absence_request_no_note() {
-        let json =
-            r#"{"absence_type":"vacation","start_date":"2026-04-01","end_date":"2026-04-05"}"#;
-        let req: CreateAbsenceRequest = serde_json::from_str(json).unwrap();
-        assert!(req.note.is_none());
-    }
-
-    // ─── CreateAnnouncementRequest ─────────────────────────────────────
-
-    #[test]
-    fn test_create_announcement_request_full() {
-        let json = r#"{
-            "title":"Maintenance",
-            "message":"Lot A closed on Monday",
-            "severity":"warning",
-            "active":true,
-            "expires_at":"2026-04-01T00:00:00Z"
-        }"#;
-        let req: CreateAnnouncementRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.title, "Maintenance");
-        assert_eq!(req.message, "Lot A closed on Monday");
-        assert_eq!(req.active, Some(true));
-        assert!(req.expires_at.is_some());
-    }
-
-    #[test]
-    fn test_create_announcement_request_minimal() {
-        let json = r#"{"title":"Info","message":"Welcome!","severity":"info"}"#;
-        let req: CreateAnnouncementRequest = serde_json::from_str(json).unwrap();
-        assert!(req.active.is_none());
-        assert!(req.expires_at.is_none());
-    }
-
     // ─── UpdatePreferencesRequest ──────────────────────────────────────
 
     #[test]
@@ -8116,53 +5349,5 @@ mod tests {
         assert_eq!(req.notifications_enabled, Some(false));
         assert!(req.email_reminders.is_none());
         assert!(req.default_duration_minutes.is_none());
-    }
-
-    // ─── CalendarQuery / CalendarEvent ─────────────────────────────────
-
-    #[test]
-    fn test_calendar_query_deserialize() {
-        let json = r#"{"from":"2026-03-01","to":"2026-03-31"}"#;
-        let q: CalendarQuery = serde_json::from_str(json).unwrap();
-        assert_eq!(q.from.as_deref(), Some("2026-03-01"));
-        assert_eq!(q.to.as_deref(), Some("2026-03-31"));
-    }
-
-    #[test]
-    fn test_calendar_event_serialize_skip_none() {
-        let event = CalendarEvent {
-            id: "evt-1".to_string(),
-            event_type: "booking".to_string(),
-            title: "Slot A3".to_string(),
-            start: Utc::now(),
-            end: Utc::now() + TimeDelta::hours(2),
-            lot_name: None,
-            slot_number: None,
-            status: None,
-        };
-        let json = serde_json::to_string(&event).unwrap();
-        assert!(!json.contains("lot_name"));
-        assert!(!json.contains("slot_number"));
-        assert!(!json.contains("status"));
-    }
-
-    #[test]
-    fn test_calendar_event_serialize_with_optionals() {
-        let event = CalendarEvent {
-            id: "evt-2".to_string(),
-            event_type: "booking".to_string(),
-            title: "Slot B1".to_string(),
-            start: Utc::now(),
-            end: Utc::now() + TimeDelta::hours(1),
-            lot_name: Some("Lot Alpha".to_string()),
-            slot_number: Some(42),
-            status: Some("confirmed".to_string()),
-        };
-        let json = serde_json::to_string(&event).unwrap();
-        assert!(json.contains("Lot Alpha"));
-        assert!(json.contains("42"));
-        assert!(json.contains("confirmed"));
-        // Check rename
-        assert!(json.contains(r#""type":"booking"#));
     }
 }
